feat: (domain, data) update field model and entities

+++ lib/domain/entities/tv/seasons.dart
@@ -8,7 +8,7 @@ class Seasons extends Equatable {
   String? overview;
   String? posterPath;
   int? seasonNumber;
-  int? voteAverage;
+  double? voteAverage;
 
   Seasons(
       {required this.airDate,
